Adicionando header e atualizando strlen

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: hnass-pe <hnass-pe@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/17 16:38:41 by hnass-pe          #+#    #+#             */
-/*   Updated: 2021/08/17 16:49:49 by hnass-pe         ###   ########.fr       */
+/*   Updated: 2021/08/17 17:12:34 by hnass-pe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	cont;
